@@ -31,10 +31,10 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  intl: ^0.17.0
+  intl: ^0.18.0
   geocoder: ^0.2.1
-  scoped_model: ^1.1.0
-  geolocator: ^7.7.1
+  scoped_model: ^2.0.0
+  geolocator: ^9.0.2
   paging: ^0.1.0
   loadmore: ^2.0.1
   futuristic: ^0.1.3
@@ -42,52 +42,50 @@ dependencies:
 #  permission_handler: ^8.2.6
   #  cupertino_icons: ^0.1.2
   google_fonts:
-  http: ^0.13.4
+  http: ^0.13.5
 
 
-  convert: ^3.0.1
+  convert: ^3.1.1
   #  geolocator: ^7.7.1
-  toast: ^0.1.5
+  toast: ^0.3.0
   #  shared_preferences: ^2.0.8
-  shared_preferences: 0.5.6
+  shared_preferences: ^2.0.15
 
-  firebase_messaging: ^11.1.0
-  email_validator: ^2.0.1
+  firebase_messaging: ^14.2.0
+  email_validator: ^2.1.17
 
-  google_maps_flutter: ^2.1.0
-  url_launcher: ^6.0.12
+  google_maps_flutter: ^2.2.2
+  url_launcher: ^6.1.7
 
   #  geocoder: ^0.2.1
   masked_text_input_formatter: ^0.0.1
-  cached_network_image: ^3.1.0
-  image_picker: ^0.8.4+4
-  firebase_analytics: ^8.3.4
-  flutter_local_notifications: ^9.2.0
+  cached_network_image: ^3.2.3
+  image_picker: ^0.8.6
+  firebase_analytics: ^10.0.8
+  flutter_local_notifications: ^13.0.0
   #  device_info:
   #  flutter_native_admob:
   #  flutter_facebook_login:
   flutter_html:
-  #facebook_audience_network:
-  #facebook_audience_network:
-  dio: ^4.0.1
+  dio: ^4.0.6
   #  google_sign_in: ^5.2.1
-  firebase_core: ^1.10.0
+  firebase_core: ^2.4.0
 
 dependency_overrides:
-  intl: ^0.17.0
+  intl: ^0.18.0
 
-  image_picker_platform_interface: ^2.4.1
-  plugin_platform_interface: ^2.0.2
-  process: ^3.0.9
-  platform: ^3.0.2
-  file: ^5.0.7
-  archive: ^3.1.6
-  yaml: ^3.1.0
-  args: ^2.3.0
-  test_api: ^0.4.6
-
-  matcher: ^0.12.11
+  image_picker_platform_interface: ^2.6.2
+  plugin_platform_interface: ^2.1.3
+  process: ^4.2.4
+  platform: ^3.1.0
+  file: ^6.1.4
+  archive: ^3.3.5
+  yaml: ^3.1.1
+  args: ^2.3.1
+  test_api: ^0.4.17
+  matcher: ^0.12.14
 #  http: 0.12.2
+  xml: ^5.1.0
 
 
 
@@ -97,19 +95,13 @@ dev_dependencies:
 
   flutter_driver:
     sdk: flutter
-  change_app_package_name: ^1.0.0
-  flutter_app_name: ^0.1.1
-  flutter_native_splash: ^1.3.1
-  flutter_launcher_name: "^0.0.1"
 
-
-
-  flutter_launcher_icons: ^0.9.2
+  flutter_launcher_icons: ^0.11.0
 
   flutter_test:
     sdk: flutter
-  test: ^1.19.2
-  flutter_lints: ^1.0.0
+  test: ^1.22.1
+  flutter_lints: ^2.0.1
   screenshots: ^2.1.1
 flutter_app_name:
   name: "Doctor App"
