@@ -461,11 +461,13 @@ class _LoginAsDoctorState extends State<LoginAsDoctor> {
               ],
             ),
             actions: [
-              FlatButton(
+              TextButton(
                 onPressed: () {
                   Navigator.pop(context);
                 },
-                color: Theme.of(context).accentColor,
+                style: TextButton.styleFrom(
+                    foregroundColor: Theme.of(context).accentColor
+                ),
                 child: Text(
                   OK,
                   style: GoogleFonts.poppins(
